backend/requirements.txt
fastapi
uvicorn
httpx
python-jose
minio
confluent-kafka
asyncpg
sqlalchemy
databases
python-multipart
redis
openai==0.28
psycopg2-binary
google-genai>=0.1.0
python-dotenv>=1.0.0